requirements.txt
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
google-auth==2.41.0
google-cloud-core==2.4.3
google-cloud-dialogflow-v2beta1 @ file:///Users/jblakey/projects/bidiSacSample/dialogflow-v2beta1-py
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
idna==3.10
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
rsa==4.9.1
typing_extensions==4.15.0
urllib3==2.5.0
